actualizadas funciones con size_t

+++ main-libft/main_ft_bzero.c
@@ -3,17 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   main_ft_bzero.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: strojo-h <strojo-h@student.42.fr>          +#+  +:+       +#+        */
+/*   By: strojo-h <strojo-h@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 15:23:17 by strojo-h          #+#    #+#             */
-/*   Updated: 2024/09/20 16:28:06 by strojo-h         ###   ########.fr       */
+/*   Updated: 2024/09/21 13:58:36 by strojo-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h> //write
 #include <stdlib.h> //malloc y free
+#include <stddef.h> //size_t
 
-void	*ft_memset(void *s, int c, unsigned int n);
+void	*ft_memset(void *s, int c, size_t n);
 
 void	ft_bzero(void *s, unsigned int n);
 
@@ -32,9 +33,9 @@ unsigned int	ft_atoi(const char *buffer_size)
 
 int	main(int argc, char *argv[])
 {
-	char			*buffer;
-	unsigned int	n;
-	char			fill_char;
+	char	*buffer;
+	size_t	n;
+	char	fill_char;
 
 	if (argc < 3)
 	{
